backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sse-starlette>=2.1.0
python-multipart>=0.0.9
pypdf>=6.0.0
langchain-openai>=0.3.0
langchain-core>=0.3.0
langgraph>=0.3.0
langsmith>=0.1.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
openai>=1.30.0
